Add files via upload
Talking to the bot requires adding it to your server, giving it the proper permissions, and running the python script. From there, you need to tag it with an @<botname> for it to read the message. message content is scanned by the moderation endpoint and not sent if flagged. The bot saves a short term memory of the conversation history, which will be dynamically reduced to fit the message limit by eliminating the oldest messages. This is done on a user by user basis, so everyone can have private conversations in the server or in DMs.

You can delete your message history with the command @<botname> delhistory

You can ask it to generate an image by using the command @<botname> Create <image prompt>